capif-provider/requirements.txt
argh==0.31.2
arrow==1.3.0
binaryornot==0.4.4
certifi==2021.10.8
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.1
cookiecutter==2.1.1
coverage==4.5.4
cryptography==38.0.4
evolved5g==1.0.13
exceptiongroup==1.2.1
flake8==3.9.2
idna==2.10
iniconfig==2.0.0
invoke==1.6.0
Jinja2==3.1.3
jinja2-time==0.2.0
MarkupSafe==2.1.5
mccabe==0.6.1
packaging==24.0
pathtools==0.1.2
pluggy==1.5.0
pycodestyle==2.7.0
pycparser==2.22
pyflakes==2.3.1
pyOpenSSL==22.1.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.26.0
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
types-python-dateutil==2.9.0.20240316
typing-extensions==3.10.0.2
urllib3==1.26.7
watchdog==0.9.0
